requirements/gca.txt
aiofiles==24.1.0
aiohttp==3.12.15
cloudpickle==3.1.1
cutoop==0.1.0
easyocr==1.7.2
gradio==5.47.0
httplib2==0.22.0
httpx==0.27.2
h2==4.3.0
ipdb==0.13.13
langchain-core==0.3.74
langgraph==0.6.6
matplotlib==3.10.5
numpy==1.26.4
onnxruntime==1.23.2
openai==1.99.9
opencv-python==4.11.0.86
openpyxl==3.1.5
open3d==0.19.0
pandas==2.3.1
pillow==11.0.0
pyarrow==21.0.0
ray[serve]==2.48.0
rembg==2.0.67
scipy==1.16.1
shortuuid==1.0.13
tensorboard==2.20.0
tensorboardX==2.6.4
tqdm==4.67.1
transformers==4.55.2
trimesh==4.8.2
git+https://github.com/EasternJournalist/utils3d.git@3fab839f0be9931dac7c8488eb0e1600c236e183
git+https://github.com/EasternJournalist/pipeline.git@866f059d2a05cde05e4a52211ec5051fd5f276d6